Develop jsp

+++ frontend/src/pages/diary/Blogs/PostDetail.vue
@@ -1,10 +1,10 @@
 <template>
   <div class="post-detail-container">
     <div class="d-flex post-detail">
-      <div class="d-flex flex-grow-0 post-detail-contact">
+      <div class="d-flex flex-shrink-0 flex-grow-0 post-detail-contact">
         <s-contact />
       </div>
-      <div class="d-flex">
+      <div class="d-flex flex-grow-0" style="width:100%">
         <div class="d-flex flex-column" style="width:100%">
           <div class="d-flex">
             <div
@@ -37,6 +37,9 @@
             <div>{{tmp.cdate}}</div>
           </div>
           <div class="post-detail-commit-container" v-if="checkcommit">
+            <div class="d-flex justify-center">
+              <div class="d-flex flex-grow-0" style="color:#0051cb;margin-top: 8px;margin-bottom:16px;font-size:14px;font-weight:600">Add Commit list</div>
+            </div>
             <div v-if="!commitcheck">
               <div
                 class="post-detail-commit-box"
@@ -44,7 +47,10 @@
                 :key="commit.cid"
               >
                 <div class="post-detail-commit-date">Commits on {{commit.date}}</div>
-                <div class="d-flex flex-column justify-center post-detail-commit" @click="mvUrl(commit.url)">
+                <div
+                  class="d-flex flex-column justify-center post-detail-commit"
+                  @click="mvUrl(commit.url)"
+                >
                   <div class="post-detail-commit-title">{{commit.msg}}</div>
                   <div class="d-flex">
                     <div class="post-detail-commit-author">{{commit.author}}</div>
@@ -62,7 +68,10 @@
             <div v-if="commitcheck">
               <div class="post-detail-commit-box" v-for="commit in commitList" :key="commit.cid">
                 <div class="post-detail-commit-date">Commits on {{commit.date}}</div>
-                <div class="d-flex flex-column justify-center post-detail-commit" @click="mvUrl(commit.url)">
+                <div
+                  class="d-flex flex-column justify-center post-detail-commit"
+                  @click="mvUrl(commit.url)"
+                >
                   <div class="post-detail-commit-title">{{commit.msg}}</div>
                   <div class="d-flex">
                     <div class="post-detail-commit-author">{{commit.author}}</div>
@@ -78,8 +87,10 @@
               @click="commitwide()"
             >접기</div>
           </div>
-          <div class="post-detail-content">{{tmp.content}}</div>
 
+          <div class="post-detail-content" v-html="compiledMarkdown()"></div>
+
+          <!-- <div class="post-detail-content">{{tmp.content}}</div> -->
           <div class="d-flex align-center flex-grow-0 post-detail-tag-container">
             <div
               class="d-flex flex-grow-0 post-detail-tag"
@@ -100,8 +111,24 @@
             <div @click="scrap" class="post-detail-icon">
               <v-icon color="#e8a317" size="32">stars</v-icon>
             </div>
-            <div @click="grapurl()" class="post-detail-icon">
-              <v-icon color="#808080" size="32">share</v-icon>
+            <div class="post-detail-icon">
+              <v-icon color="#808080" size="32" @click.stop="dialog=true;getUrl()">share</v-icon>
+              <!-- 모달 하나 띄워 줘야 될거같다 -->
+              <v-dialog v-model="dialog" width="400">
+                <v-card>
+                  <v-card-title>공유 하기</v-card-title>
+                  <v-card-text>
+                    <v-text-field id="ShareUrl" v-model="this.urll" readonly></v-text-field>
+                  </v-card-text>
+                  <v-card-actions>
+                    <v-spacer></v-spacer>
+                    <div
+                      class="d-flex justify-center align-center flex-grow-0 s-button-blue"
+                      @click="CopyUrlToClipboard();dialog=false"
+                    >복사</div>
+                  </v-card-actions>
+                </v-card>
+              </v-dialog>
             </div>
           </div>
           <div class="post-detail-comment-container">
@@ -112,35 +139,49 @@
             >
               <div class="d-flex align-center">
                 <div class="d-flex flex-grow-0">
-                  <img class="post-detail-comment-img" :src="comment.userinfo.img == null ? '/static/images/user.png' : comment.userinfo.img" />
+                  <img
+                    class="post-detail-comment-img"
+                    :src="comment.userinfo.img == null ? '/static/images/user.png' : comment.userinfo.img"
+                  />
                 </div>
                 <div class="post-detail-comment-id">{{comment.userinfo.nickname}}</div>
               </div>
               <div class="d-flex align-center">
-                <div class="d-flex post-detail-comment-content">{{comment.content}}</div>
+                <div class="d-flex align-center post-detail-comment-content">{{comment.content}}</div>
                 <div
-                  class="d-flex justify-center align-center flex-grow-0 post-detail-red-btn"
+                  class="d-flex justify-center align-center flex-grow-0 flex-shrink-0 post-detail-red-btn"
                   style="margin-left: 8px;"
                   v-if="isWritten(comment.uid)"
                   @click="commenterase(comment.id)"
                 >삭제</div>
               </div>
               <div class="d-flex post-detail-comment-date">{{comment.cdate}}</div>
-
             </div>
             <div class="d-flex justify-center align-center post-detail-comment-box">
               <!-- v-model="commentData.content" -->
-              <v-text-field v-if="isLogin" id="post-comment-content" dense outlined hide-details 
-              placeholder='최대 길이는 100자입니다.'
+              <v-text-field
+                v-if="isLogin"
+                id="post-comment-content"
+                dense
+                outlined
+                hide-details
+                placeholder="최대 길이는 100자입니다."
               />
               <div
                 style="margin-left: 8px"
                 class="d-flex flex-grow-0 align-center justify-center post-detail-blue-btn"
                 @click="commentwrite()"
                 v-if="isLogin"
-                
               >작성</div>
-              <v-text-field v-if="!isLogin" id="post-comment-content" dense outlined hide-details placeholder="댓글은 로그인이 필요한 서비스입니다." @click="$router.push({name:'Login'})"/>
+              <v-text-field
+                v-if="!isLogin"
+                id="post-comment-content"
+                dense
+                outlined
+                hide-details
+                placeholder="댓글은 로그인이 필요한 서비스입니다."
+                @click="$router.push({name:'Login'})"
+              />
               <div
                 style="margin-left: 8px"
                 class="d-flex flex-grow-0 align-center justify-center post-detail-blue-btn"
@@ -162,8 +203,10 @@
 import SContact from '../../../component/s-contact.vue';
 import ContentSidebar from '../../../component/ContentSidebar.vue';
 import axios from 'axios';
-import moment from 'moment';  //
+import moment from 'moment';
+import marked from 'marked'; //
 
+var renderer = new marked.Renderer();
 
 export default {
   name: 'PostDetail',
@@ -187,8 +230,10 @@ export default {
       tags: {},
       commitList: [],
       comments: [],
-      isLogin:false,
-      printLikeCnt :'',
+      isLogin: false,
+      printLikeCnt: '',
+      dialog: false,
+      urll: '',
     };
   },
   async created() {
@@ -202,9 +247,8 @@ export default {
       let tmpspace = await this.$api.postdetail(this.id.pid);
       this.tmp = tmpspace;
       this.printLikeCnt = this.tmp.cntLike;
-      
     } catch (e) {
-      console.log(e);
+      // console.log(e);
     }
     //좋아요 데이터 가져오기
     try {
@@ -220,15 +264,15 @@ export default {
         this.likeData.likechecked = false;
       }
     } catch (e) {
-      console.log(e);
+      // console.log(e);
     }
     //commit data가져오기
     try {
       let tmpspace2 = await this.$api.getPostCommit(this.id.pid);
       this.commitList = tmpspace2;
-      console.log('커밋 부르기 성공');
+      // console.log('커밋 부르기 성공');
     } catch (e) {
-      console.log(e);
+      // console.log(e);
     }
 
     //tag 데이터 가져오기
@@ -236,86 +280,64 @@ export default {
       let tmpspace3 = await this.$api.tagIndex(this.id.pid);
       this.tags = tmpspace3;
     } catch (e) {
-      console.log(e);
+      // console.log(e);
     }
 
-    console.log('***' + this.commitList[0]);
+    // console.log('***' + this.commitList[0]);
   },
   methods: {
+    compiledMarkdown () {
+      if (!this.tmp || !this.tmp.content) {
+        return ''
+      }
+
+      if (this.commitList.length != 0) {
+        renderer.em = (text) => {
+          var indexNumber = text.indexOf('/');
+          if (indexNumber !== -1 && text.substr(indexNumber - 1, 1) !== '\\') {
+            var idx = text.substr(indexNumber + 1);
+            var commit = this.commitList[idx - 1];
+            var res =
+              '<div class="post-detail-commit-container" ><div class="contents-commit-box">' +
+              `<div class="d-flex flex-column justify-center contents-commit" ><div class="contents-commit-title"> #` +
+              idx +
+              ' ' +
+              commit.msg +
+              '</div><div class="d-flex"> <div class="contents-commit-author">' +
+              commit.author +
+              ' committed on ' +
+              commit.date +
+              '</div></div></div></div></div>';
+            return res;
+          }
+          return '<em>' + text.replace('\\/', '/') + '</em>';
+        };
+      }
+
+      return marked(this.tmp.content, { renderer: renderer });
+    },
     async getComment() {
       try {
         let temp = await this.$api.getCommentlist(this.id.pid);
         this.comments = temp;
-        console.log(temp,'vdasdfsafbvfad')
+        // console.log(temp, 'vdasdfsafbvfad');
         // console.log('Comment 목록 가져오기 성공')
       } catch (e) {
-        console.log(e);
+        // console.log(e);
       }
     },
-    async created(){
-        this.uid=this.$store.state.user.id
-        //comment 데이터 가져오기
-        this.getComment();
-
-        //post 데이터 가져오기
-        try{
-            let tmpspace = await this.$api.postdetail(this.id.pid)
-            this.tmp =tmpspace
-            // console.log('성공')
-        }catch(e){
-            console.log(e)
-        }
-        //좋아요 데이터 가져오기
-        try{
-            let tmpspace1= await this.$api.likedatas(this.id.pid)
-                this.likedummy = tmpspace1
-            if (tmpspace1.length!=0){
-
-                if(tmpspace1.status==1){
-                    this.likeData.likechecked=true
-                }else{
-                    this.likeData.likechecked=false
-                }}
-            else{
-                this.likeData.likechecked=false
-            }
-        }catch(e){
-            console.log(e)
-        }
-        //commit data가져오기
-        try{
-            let tmpspace2 = await this.$api.getPostCommit(this.id.pid)
-            this.commitList= tmpspace2
-            console.log('커밋 부르기 성공')
-            
-
-        }catch(e){
-            console.log(e)
-        }
-
-    //   var ymd = d.toISOString().substr(0, 10);
-    //   var timestamp =
-    //     ('00' + d.getHours()).slice(-2) +
-    //     ':' +
-    //     ('00' + d.getMinutes()).slice(-2) +
-    //     ':' +
-    //     ('00' + d.getSeconds()).slice(-2);
-
-    //   this.commentData.cDate = ymd + ' ' + timestamp;
-      //console.log(this.commentData.cDate)
-    },
     goback() {
-      this.$router.go(-1);
+      this.$router.push({ name: 'BlogDetail', params: { did: this.tmp.did } });
     },
     async like() {
-      if(this.isLogin){
+      if (this.isLogin) {
         if (this.likeData.likechecked == true) {
           this.likeData.likechecked = false;
           this.$api.likeDislike({ pid: this.id.pid, status: 0 });
           this.printLikeCnt -= 1;
-          console.log('좋아요 취소');
+          // console.log('좋아요 취소');
         } else {
-          console.log('좋아요');
+          // console.log('좋아요');
           this.likeData.likechecked = true;
           this.$api.likeDislike({ pid: this.id.pid, status: 1 });
           this.printLikeCnt += 1;
@@ -326,16 +348,19 @@ export default {
           this.tmp = tmpspace;
           // console.log('성공')
         } catch (e) {
-          console.log(e);
+          // console.log(e);
         }
-        this.$forceUpdate();}
-        else{
-          alert("로그인을 먼저 해주세요")
-        }
+        this.$forceUpdate();
+      } else {
+        alert('로그인을 먼저 해주세요');
+      }
     },
     searchTag(tagName) {
       document.getElementById('header-text').value = tagName;
-      this.$router.push({ name: 'tmp', params: { key: tagName, type: 'title' } });
+      this.$router.push({
+        name: 'tmp',
+        params: { key: tagName, type: 'tag' },
+      });
     },
     grapurl() {
       alert(this.$route.path);
@@ -348,34 +373,31 @@ export default {
       }
     },
     scrap() {
-      if(this.isLogin){
-      this.$api.makeScrap({ pid: this.id.pid, status: 1 });
-      alert('스크랩 되었습니다');}
-      else{
-        alert('로그인을 먼저 해주세요')
+      if (this.isLogin) {
+        this.$api.makeScrap({ pid: this.id.pid, status: 1 });
+        alert('스크랩 되었습니다');
+      } else {
+        alert('로그인을 먼저 해주세요');
       }
     },
     async commentwrite() {
-      
-
       this.commentData.content = document.getElementById(
         'post-comment-content'
       ).value;
-      if (this.commentData.content =="" || this.commentData.content==null){
-        alert('빈 댓글은 허용하지 않습니다')
-      }else if(this.commentData.content.length >100 ){
-        alert('댓글 최대 허용 길이는 100자 입니다')
-        document.getElementById(
-        'post-comment-content'
-        ).value = ""
+      if (this.commentData.content == '' || this.commentData.content == null) {
+        alert('빈 댓글은 허용하지 않습니다');
+      } else if (this.commentData.content.length > 100) {
+        alert('댓글 최대 허용 길이는 100자 입니다');
+        document.getElementById('post-comment-content').value = '';
+      } else {
+        //alert('준비중입니다.')
+        this.commentData.uid = this.$store.state.user.id;
+        // this.getcDate()
+        this.commentData.pid = this.id.pid;
+        await this.$api.createComment(this.commentData);
+        document.getElementById('post-comment-content').value = '';
+        this.getComment();
       }
-      else{
-      //alert('준비중입니다.')
-      this.commentData.uid = this.$store.state.user.id;
-      // this.getcDate()
-      this.commentData.pid = this.id.pid;
-      await this.$api.createComment(this.commentData);
-      this.getComment();}
     },
     async commenterase(commentid) {
       // console.log("CKCK commentid")
@@ -394,17 +416,39 @@ export default {
       }
     },
     async deleteP(postid) {
-      try {
-        this.$api.deletePost(postid);
-        // console.log('성공')
-        this.$router.go(-1);
-      } catch (e) {
-        console.log(e);
+      if (confirm('정말 삭제하시겠습니까 ?') == true) {
+        try {
+          this.$api.deletePost(postid);
+          // console.log('성공')
+          this.$router.push({
+            name: 'BlogDetail',
+            params: { did: this.tmp.did },
+          });
+        } catch (e) {
+          // console.log(e);
+        }
+      } else {
+        return;
       }
     },
-    mvUrl(url){
-      window.open(url, "_blank");
-    }
+    mvUrl(url) {
+      window.open(url, '_blank');
+    },
+    CopyUrlToClipboard() {
+      var obShareUrl = document.getElementById('ShareUrl');
+      obShareUrl.value = window.document.location.href; // 현재 URL 을 세팅해 줍니다.
+
+      obShareUrl.select(); // 해당 값이 선택되도록 select() 합니다
+      document.execCommand('copy'); // 클립보드에 복사합니다.
+
+      obShareUrl.blur(); // 선택된 것을 다시 선택안된것으로 바꿈니다.
+      alert('URL이 클립보드에 복사되었습니다');
+    },
+    getUrl() {
+      var obShareUrl = document.getElementById('ShareUrl');
+      // obShareUrl.value = window.document.location.href;
+      this.urll = window.document.location.href;
+    },
   },
   computed: {
     limited: function () {
@@ -420,7 +464,7 @@ export default {
     likechecking() {
       if (this.likedummy.status == 1) {
         this.likeData.likechecked == true;
-      } else if ((this.likedummy, length == 0)) {
+      } else if ((this.likedummy.length == 0)) {
         this.likeData.likechecked == false;
       } else {
         this.likeData.likechecked == false;
@@ -523,10 +567,10 @@ export default {
 }
 
 .post-detail-like {
-    color: #db4455;
-    font-weight: bold;
-    font-size: 20px;
-    margin-right: 8px;
+  color: #db4455;
+  font-weight: bold;
+  font-size: 20px;
+  margin-right: 8px;
 }
 
 .post-detail-commit-container {
@@ -611,6 +655,7 @@ export default {
 }
 
 .post-detail-comment-date {
+  margin-top: 4px;
   color: rgb(102, 102, 102);
   font-size: 12px;
 }
@@ -621,6 +666,47 @@ export default {
 }
 
 .post-detail-comment-content {
+  min-height: 40px;
   font-size: 14px;
+}
+
+.contents-commit-box {
+  border-left: solid 2px #dde3ea;
+  padding-left: 16px;
+  margin-bottom: 22px;
+}
+
+.contents-commit {
+  border: solid 1px #dde3ea;
+  padding: 8px;
+  height: 60px;
+  border-radius: 6px;
+}
+
+.contents-commit:hover {
+  background: #0051cb11;
+  cursor: pointer;
+}
+
+.contents-commit-title {
+  font-size: 14px;
+  font-weight: 800;
+}
+
+.contents-commit-date {
+  margin-bottom: 12px;
+  font-size: 12px;
+  font-weight: normal;
+}
+
+.contents-commit-author {
+  font-size: 12px;
+  font-weight: 600;
+  color: #24292e;
+}
+
+.contents-commit-sha {
+  font-size: 10px;
+  font-weight: normal;
 }
 </style>
